+++ src/PairedReader.cpp
@@ -1,9 +1,34 @@
 
 #include <thread>
 #include "PairedReader.h"
+#include "FastqReader.h"
+#include "BoostZipReader.h"
 
-PairedReader::PairedReader(string fastqgz_1, string fastqgz_2) :
-		Reader1(fastqgz_1), Reader2(fastqgz_2) {
+PairedReader::PairedReader(string fastqgz_1, string fastqgz_2) {
+
+	if (fastqgz_1.length() > 3){
+		string suffix = fastqgz_1.substr(fastqgz_1.length()-3);
+		if (suffix.compare(".gz") == 0){
+			// use zip reader
+			Reader1 = new BoostZipReader(fastqgz_1);
+
+		}
+		else{
+			// use fastq reader
+			Reader1 = new FastqReader(fastqgz_1);
+		}
+	}
+	if (fastqgz_2.length() > 3){
+		string suffix = fastqgz_2.substr(fastqgz_2.length()-3);
+		if (suffix.compare(".gz") == 0){
+			// use zip reader
+			Reader2 = new BoostZipReader(fastqgz_2);
+		}
+		else{
+			// use fastq reader
+			Reader2 = new FastqReader(fastqgz_2);
+		}
+	}
 }
 
 std::vector<std::pair<fq_read*, fq_read*>>* PairedReader::next_reads(
@@ -13,8 +38,8 @@ std::vector<std::pair<fq_read*, fq_read*>>* PairedReader::next_reads(
 
 	fq_read *read1, *read2;
 	for (size_t i = 0; i < max_size; i++) {
-		read1 = Reader1.next_read();
-		read2 = Reader2.next_read();
+		read1 = Reader1->next_read();
+		read2 = Reader2->next_read();
 		if (read1 != NULL && read2 != NULL) {
 			pairs->push_back(std::pair<fq_read*, fq_read*>(read1, read2));
 		}
@@ -26,7 +51,7 @@ std::vector<std::pair<fq_read*, fq_read*>>* PairedReader::next_reads(
 	return pairs;
 }
 
-void fill_reads(ZipFastqReader* reader, size_t max_size, fq_read ***reads) {
+void fill_reads(IFastqReader* reader, size_t max_size, fq_read ***reads) {
 	for (size_t i = 0; i < max_size; i++) {
 		fq_read *tmp = reader->next_read();
 		if (tmp != NULL){
@@ -48,14 +73,14 @@ std::vector<std::pair<fq_read*, fq_read*>>* PairedReader::next_reads2(
 	fq_read** reads2 = new fq_read*[max_size];
 
 	if (reading_threads == 2){
-		std::thread to1(fill_reads,  &Reader1, max_size, &reads1);
-		std::thread to2(fill_reads,  &Reader2, max_size, &reads2);
+		std::thread to1(fill_reads,  Reader1, max_size, &reads1);
+		std::thread to2(fill_reads,  Reader2, max_size, &reads2);
 		to1.join();
 		to2.join();
 	}
 	else{
-		fill_reads(&Reader1, max_size, &reads1);
-		fill_reads(&Reader2, max_size, &reads2);
+		fill_reads(Reader1, max_size, &reads1);
+		fill_reads(Reader2, max_size, &reads2);
 	}
 
 	for(size_t i = 0; i < max_size; i++){
@@ -73,5 +98,7 @@ std::vector<std::pair<fq_read*, fq_read*>>* PairedReader::next_reads2(
 }
 
 PairedReader::~PairedReader() {
+	delete Reader1;
+	delete Reader2;
 }