+++ src/idemuxCPP_cmdl.c
@@ -41,12 +41,13 @@ const char *idemuxCPP_args_info_help[] = {
   "  -o, --out=STRING              Where to write the output files.\n                                    (default=`./')",
   "  -s, --sample-sheet=STRING     Outputs a csv file containing sample names, i7,\n                                  i5 and i1 barcodes.\n                                    (default=`sample-sheet.csv')",
   "  -b, --barcode-corrections=STRING\n                                Outputs a csv file that contains the number of\n                                  corrected barcodes",
-  "  -5, --i5-rc                   et this flag if the i5 barcode has been\n                                  sequenced as reverse complement and the\n                                  barcodes you provided should be reverse\n                                  complemented.\n                                    (default=off)",
-  "  -i, --i1-start=INT            Start position of the i1 index (0-based) on\n                                  read 2.\n                                    (default=`10')",
+  "  -5, --i5-rc                   Set this flag if the i5 barcode has been\n                                  sequenced as reverse complement and the\n                                  barcodes you provided should be reverse\n                                  complemented.\n                                    (default=off)",
+  "  -i, --i1-start=INT            Start position of the i1 index (1-based) on\n                                  read 2.\n                                    (default=`11')",
   "  -q, --queue-size=INT          Queue size for reads that will be processed in\n                                  one block.\n                                    (default=`4000000')",
   "  -r, --reading-threads=INT     Number of threads used for reading gz files.\n                                  Either 1 or 2 (one thread per input file is\n                                  used).\n                                    (default=`2')",
   "  -w, --writing-threads=INT     Number of threads used for writing gz files.\n                                  Default is the number of processor cores.\n",
   "  -p, --processing-threads=INT  Number of threads used for processing the error\n                                  correction. Default is the number of\n                                  processor cores.\n",
+  "  -d, --demux-only              Do a one on one mapping for the barcodes\n                                  specified in the sample sheet. No error\n                                  correction will be done. Barcodes that do not\n                                  match are written to the undetermined reads\n                                  file.  (default=off)",
   "  -v, --verbose                 Verbose.\n                                    (default=off)",
     0
 };
@@ -88,6 +89,7 @@ void clear_given (struct idemuxCPP_args_info *args_info)
   args_info->reading_threads_given = 0 ;
   args_info->writing_threads_given = 0 ;
   args_info->processing_threads_given = 0 ;
+  args_info->demux_only_given = 0 ;
   args_info->verbose_given = 0 ;
 }
 
@@ -106,7 +108,7 @@ void clear_args (struct idemuxCPP_args_info *args_info)
   args_info->barcode_corrections_arg = NULL;
   args_info->barcode_corrections_orig = NULL;
   args_info->i5_rc_flag = 0;
-  args_info->i1_start_arg = 10;
+  args_info->i1_start_arg = 11;
   args_info->i1_start_orig = NULL;
   args_info->queue_size_arg = 4000000;
   args_info->queue_size_orig = NULL;
@@ -114,6 +116,7 @@ void clear_args (struct idemuxCPP_args_info *args_info)
   args_info->reading_threads_orig = NULL;
   args_info->writing_threads_orig = NULL;
   args_info->processing_threads_orig = NULL;
+  args_info->demux_only_flag = 0;
   args_info->verbose_flag = 0;
   
 }
@@ -136,7 +139,8 @@ void init_args_info(struct idemuxCPP_args_info *args_info)
   args_info->reading_threads_help = idemuxCPP_args_info_help[11] ;
   args_info->writing_threads_help = idemuxCPP_args_info_help[12] ;
   args_info->processing_threads_help = idemuxCPP_args_info_help[13] ;
-  args_info->verbose_help = idemuxCPP_args_info_help[14] ;
+  args_info->demux_only_help = idemuxCPP_args_info_help[14] ;
+  args_info->verbose_help = idemuxCPP_args_info_help[15] ;
   
 }
 
@@ -291,6 +295,8 @@ idemuxCPP_cmdline_parser_dump(FILE *outfile, struct idemuxCPP_args_info *args_in
     write_into_file(outfile, "writing-threads", args_info->writing_threads_orig, 0);
   if (args_info->processing_threads_given)
     write_into_file(outfile, "processing-threads", args_info->processing_threads_orig, 0);
+  if (args_info->demux_only_given)
+    write_into_file(outfile, "demux-only", 0, 0 );
   if (args_info->verbose_given)
     write_into_file(outfile, "verbose", 0, 0 );
   
@@ -1191,6 +1197,7 @@ idemuxCPP_cmdline_parser_internal (
         { "reading-threads",	1, NULL, 'r' },
         { "writing-threads",	1, NULL, 'w' },
         { "processing-threads",	1, NULL, 'p' },
+        { "demux-only",	0, NULL, 'd' },
         { "verbose",	0, NULL, 'v' },
         { 0,  0, 0, 0 }
       };
@@ -1200,7 +1207,7 @@ idemuxCPP_cmdline_parser_internal (
       custom_opterr = opterr;
       custom_optopt = optopt;
 
-      c = custom_getopt_long (argc, argv, "hV1:2:o:s:b:5i:q:r:w:p:v", long_options, &option_index);
+      c = custom_getopt_long (argc, argv, "hV1:2:o:s:b:5i:q:r:w:p:dv", long_options, &option_index);
 
       optarg = custom_optarg;
       optind = custom_optind;
@@ -1285,7 +1292,7 @@ idemuxCPP_cmdline_parser_internal (
             goto failure;
         
           break;
-        case '5':	/* et this flag if the i5 barcode has been sequenced as reverse complement and the barcodes you provided should be reverse complemented.
+        case '5':	/* Set this flag if the i5 barcode has been sequenced as reverse complement and the barcodes you provided should be reverse complemented.
 .  */
         
         
@@ -1296,13 +1303,13 @@ idemuxCPP_cmdline_parser_internal (
             goto failure;
         
           break;
-        case 'i':	/* Start position of the i1 index (0-based) on read 2.
+        case 'i':	/* Start position of the i1 index (1-based) on read 2.
 .  */
         
         
           if (update_arg( (void *)&(args_info->i1_start_arg), 
                &(args_info->i1_start_orig), &(args_info->i1_start_given),
-              &(local_args_info.i1_start_given), optarg, 0, "10", ARG_INT,
+              &(local_args_info.i1_start_given), optarg, 0, "11", ARG_INT,
               check_ambiguity, override, 0, 0,
               "i1-start", 'i',
               additional_error))
@@ -1357,6 +1364,16 @@ idemuxCPP_cmdline_parser_internal (
               &(local_args_info.processing_threads_given), optarg, 0, 0, ARG_INT,
               check_ambiguity, override, 0, 0,
               "processing-threads", 'p',
+              additional_error))
+            goto failure;
+        
+          break;
+        case 'd':	/* Do a one on one mapping for the barcodes specified in the sample sheet. No error correction will be done. Barcodes that do not match are written to the undetermined reads file..  */
+        
+        
+          if (update_arg((void *)&(args_info->demux_only_flag), 0, &(args_info->demux_only_given),
+              &(local_args_info.demux_only_given), optarg, 0, 0, ARG_FLAG,
+              check_ambiguity, override, 1, 0, "demux-only", 'd',
               additional_error))
             goto failure;