+++ CHANGELOG.rst
@@ -2,7 +2,18 @@
 History
 =======
 
-0.1.0 (2020-09-010)
+0.1.3 (2020-11-03)
+------------------
+
+* bugfix zip file reader
+* optional correction count output file
+
+0.1.2 (2020-09-10)
+------------------
+
+* Second release.
+
+0.1.0 (2020-09-10)
 ------------------
 
 * First release.